refactor(py3): remove unnecessary float casts in division

+++ emission/analysis/intake/segmentation/trip_segmentation_methods/dwell_segmentation_dist_filter.py
@@ -94,7 +94,7 @@ class DwellSegmentationDistFilter(eaist.TripSegmentationMethod):
 
         # --- Define thresholds for segmentation ---
         # For a candidate segmentation, if the speed is below this threshold, then the phone may be stationary.
-        speed_threshold = float(self.distance_threshold * 2) / (self.time_threshold / 2)
+        speed_threshold = (self.distance_threshold * 2) / (self.time_threshold / 2)
 
         with ect.Timer() as t_loop:
             # --- Identify candidate segmentation points ---
